nshuti/requirements.txt
asgiref==3.6.0
boto3==1.26.115
botocore==1.29.115
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==1.3.0
Django==4.2
django-storages==1.13.2
django-tinymce4-lite==1.8.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
jsmin==3.0.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
